Рефакторинг Логика открытия таба для выбора пользователя вынесена во ViewModel Удаление пользователя вынесено во ViewModel
